bruin/biomass/assets/ingestion/requirements.txt
requests
google-cloud-storage
urllib3
pandas
openpyxl
xlrd
